Improve some things about keymode handling

+++ src/resource_managers/ChartFactory.cpp
@@ -488,7 +488,7 @@ getComponentsForPlayer(const ChartFactory::PlayerSpecificData& player,
       chartData.getLength(),
       chartData.getSha256(),
       chartData.getMd5(),
-      static_cast<int>(keymode),
+      keymode,
       player.replayedScore != nullptr
         ? player.replayedScore->getResult()->getUnixTimestamp()
         : 0,
